+++ app/Site.php
@@ -26,11 +26,23 @@ class Site extends Model
         'type',
         'source_repo',
         'source_branch',
-        'document_root',
+        'project_root',
+        'public_dir',
         'redirect_type',
         'redirect_to',
         'is_enabled',
     ];
+
+    public function getDocumentRootAttribute(): string
+    {
+        $docroot = $this->project_root;
+
+        if ('laravel' === $this->type) {
+            $docroot .= $this->public_dir;
+        }
+
+        return $docroot;
+    }
 
     public function getLogsAttribute(): array
     {
@@ -50,9 +62,11 @@ class Site extends Model
 
     public function readLog(string $log): string
     {
-        $path = Str::startsWith($this->logs[$log]['path'], '/') ? '' : (
-            Str::beforeLast($this->document_root, '/public') . '/'
-        ) . $this->logs[$log]['path'];
+        $path = $this->logs[$log]['path'];
+
+        if (!Str::startsWith($path, '/')) {
+            $path = $this->project_root . '/' . $path;
+        }
 
         exec('sudo cat ' . escapeshellarg($path), $file);