+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 * An install script! Installation is now the same across dev *and* "prod"
+* Public folder can be defined separately from docroot for Laravel projects
 
 ### Changed
-* Project folder paths will now fall back to site name when domain isn't set
+* Project folder paths will fall back to site name when domain isn't set
 
 ### Fixed
 * Nginx sites no longer try to use the old PHP 7.3 socket
+* Overflowing dropdowns in site editor don't overflow past the grid any more
 
 
 ## [0.9.0] - 2020-08-11